# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "urtorrentd",
    "description": "HTTP API of urtorrentd, a BitTorrent daemon on the urtorrent library. Units: bytes, bytes per second, seconds, unix seconds; `null` means unknown or unlimited. Errors always have the `ErrorBody` shape with a stable `code`.",
    "contact": {
      "name": "urtorrentd contributors"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/app": {
      "get": {
        "tags": [
          "app"
        ],
        "summary": "About the daemon.",
        "operationId": "get_app_info",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppInfo"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/app/shutdown": {
      "post": {
        "tags": [
          "app"
        ],
        "summary": "Shut the daemon down gracefully (trackers are told, state is saved).",
        "operationId": "shutdown",
        "responses": {
          "202": {
            "description": "Shutting down."
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/fs/directory": {
      "get": {
        "tags": [
          "app"
        ],
        "summary": "List a directory on the daemon's machine (for choosing save paths).",
        "operationId": "list_directory",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "description": "Absolute directory.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "description": "What to include.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/DirectoryMode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DirectoryEntry"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/settings": {
      "get": {
        "tags": [
          "app"
        ],
        "summary": "The settings.",
        "operationId": "get_settings",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "app"
        ],
        "summary": "Change settings. Only the fields present change; the result is the\ncomplete settings. Live fields apply at once; the engine-tuning ones after\na restart (see `GET /app`, `restart_required`).",
        "operationId": "patch_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppInfo": {
        "type": "object",
        "description": "About the daemon.",
        "required": [
          "version",
          "api_version",
          "library",
          "pid",
          "started_at",
          "data_dir",
          "default_save_path",
          "listen_port",
          "restart_required"
        ],
        "properties": {
          "version": {
            "type": "string",
            "description": "Daemon version."
          },
          "api_version": {
            "type": "string",
            "description": "API version (the base path's)."
          },
          "library": {
            "type": "string",
            "description": "The torrent library and its version."
          },
          "pid": {
            "type": "integer",
            "format": "int32",
            "description": "Process id.",
            "minimum": 0
          },
          "started_at": {
            "type": "integer",
            "format": "int64",
            "description": "Start time, unix seconds.",
            "minimum": 0
          },
          "data_dir": {
            "type": "string",
            "description": "The data directory."
          },
          "default_save_path": {
            "type": "string",
            "description": "The default save path."
          },
          "listen_port": {
            "type": "integer",
            "format": "int32",
            "description": "The peer listen port in use.",
            "minimum": 0
          },
          "restart_required": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Settings changed since the start that apply only after a restart."
          }
        }
      },
...
      "DirectoryEntry": {
        "type": "object",
        "description": "A directory entry.",
        "required": [
          "name",
          "path",
          "is_dir"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Name."
          },
          "path": {
            "type": "string",
            "description": "Absolute path."
          },
          "is_dir": {
            "type": "boolean",
            "description": "A directory."
          }
        }
      },
      "DirectoryMode": {
        "type": "string",
        "description": "What a directory listing includes.",
        "enum": [
          "all",
          "dirs",
          "files"
        ]
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "Every error response has this body.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "The error."
          }
        }
      },
...
      "ErrorDetail": {
        "type": "object",
        "description": "The error object.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Stable code."
          },
          "message": {
            "type": "string",
            "description": "Human-readable explanation."
          }
        }
      },
...
      "Settings": {
        "type": "object",
        "description": "The daemon's settings. Every field is always present.",
        "required": [
          "listen_port",
          "random_port",
          "listen_v4",
          "listen_v6",
          "identity",
          "encryption",
          "transports",
          "dht",
          "pex",
          "lsd",
          "dht_bootstrap_nodes",
          "max_connections",
          "max_connections_per_torrent",
          "max_uploads",
          "max_uploads_per_torrent",
          "banned_ips",
          "download_limit",
          "upload_limit",
          "alt_download_limit",
          "alt_upload_limit",
          "alt_speed_enabled",
          "queueing_enabled",
          "max_active_downloads",
          "max_active_uploads",
          "max_active_torrents",
          "count_slow_torrents",
          "save_path",
          "download_path",
          "add_stopped",
          "add_to_top_of_queue",
          "preallocate",
          "auto_management",
          "add_trackers",
          "max_ratio",
          "max_seeding_time",
          "max_inactive_seeding_time",
          "share_limit_action",
          "hash_threads",
          "max_open_files",
          "max_checking",
          "piece_extent_affinity",
          "max_concurrent_announces",
          "disk_thread",
          "zero_copy_send",
          "api_session_timeout",
          "api_max_auth_failures",
          "api_ban_duration",
          "api_bypass_local_auth",
          "api_auth_whitelist",
          "api_allowed_hosts",
          "api_csrf_protection"
        ],
        "properties": {
          "listen_port": {
            "type": "integer",
            "format": "int32",
            "description": "TCP and UDP listen port for peers (0 = pick an ephemeral port at start).",
            "default": 6881,
            "minimum": 0
          },
          "random_port": {
            "type": "boolean",
            "description": "Pick a new random listen port at every start.",
            "default": false
          },
          "listen_v4": {
            "type": [
              "string",
              "null"
            ],
            "description": "IPv4 listen address; `null` turns IPv4 off (no socket, no IPv4 peers).",
            "default": "0.0.0.0"
          },
          "listen_v6": {
            "type": [
              "string",
              "null"
            ],
            "description": "IPv6 listen address; `null` turns IPv6 off.",
            "default": "::"
          },
          "identity": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Identity",
                "description": "The identity peers and trackers see."
              }
            ],
            "default": "native"
          },
          "encryption": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Encryption",
                "description": "Peer connection encryption (MSE)."
              }
            ],
            "default": "enabled"
          },
          "transports": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Transports",
                "description": "Peer transports (TCP, uTP) and their order."
              }
            ],
            "default": "prefer_tcp"
          },
          "dht": {
            "type": "boolean",
            "description": "The DHT (BEP 5). Private torrents never use it.",
            "default": true
          },
          "pex": {
            "type": "boolean",
            "description": "Peer exchange (BEP 11). Private torrents never use it.",
            "default": true
          },
          "lsd": {
            "type": "boolean",
            "description": "Local service discovery (BEP 14). Private torrents never use it.",
            "default": true
          },
          "dht_bootstrap_nodes": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "DHT bootstrap routers (`host:port`); `null` = the identity's defaults.\nApplies after a restart.",
            "default": null
          },
          "max_connections": {
            "type": "integer",
            "format": "int32",
            "description": "Peer connections across all torrents.",
            "default": 500,
            "minimum": 0
          },
          "max_connections_per_torrent": {
            "type": "integer",
            "format": "int32",
            "description": "Peer connections per torrent (the default for torrents without their own cap).",
            "default": 50,
            "minimum": 0
          },
          "max_uploads": {
            "type": "integer",
            "format": "int32",
            "description": "Upload slots across all torrents.",
            "default": 8,
            "minimum": 0
          },
          "max_uploads_per_torrent": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Upload slots per torrent for new torrents; `null` = only the global budget.",
            "default": null,
            "minimum": 0
          },
          "banned_ips": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Addresses banned from every torrent.",
            "default": []
          },
          "download_limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Global download limit in bytes per second; `null` = unlimited.",
            "default": null,
            "minimum": 0
          },
          "upload_limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Global upload limit in bytes per second; `null` = unlimited.",
            "default": null,
            "minimum": 0
          },
          "alt_download_limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Alternative download limit (bytes per second) used while `alt_speed_enabled`.",
            "default": 10485760,
            "minimum": 0
          },
          "alt_upload_limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Alternative upload limit (bytes per second) used while `alt_speed_enabled`.",
            "default": 10485760,
            "minimum": 0
          },
          "alt_speed_enabled": {
            "type": "boolean",
            "description": "Use the alternative limits.",
            "default": false
          },
          "queueing_enabled": {
            "type": "boolean",
            "description": "Limit how many torrents are active at once (the queue).",
            "default": false
          },
          "max_active_downloads": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Active downloading torrents while queueing; `null` = unlimited.",
            "default": 3,
            "minimum": 0
          },
          "max_active_uploads": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Active seeding torrents while queueing; `null` = unlimited.",
            "default": 3,
            "minimum": 0
          },
          "max_active_torrents": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Active torrents in total while queueing; `null` = unlimited.",
            "default": 5,
            "minimum": 0
          },
          "count_slow_torrents": {
            "type": "boolean",
            "description": "Count torrents that move no data (below 2 KiB/s for 60 s) towards the queue limits.",
            "default": false
          },
          "save_path": {
            "type": "string",
            "description": "Where new torrents are saved (absolute path).",
            "default": ""
          },
          "download_path": {
            "type": [
              "string",
              "null"
            ],
            "description": "Where incomplete torrents are downloaded before moving to their save\npath (absolute path); `null` = straight to the save path.",
            "default": null
          },
          "add_stopped": {
            "type": "boolean",
            "description": "Add new torrents stopped.",
            "default": false
          },
          "add_to_top_of_queue": {
            "type": "boolean",
            "description": "Put new torrents at the front of the queue.",
            "default": false
          },
          "preallocate": {
            "type": "boolean",
            "description": "Allocate content files at full size when they are created.",
            "default": false
          },
          "auto_management": {
            "type": "boolean",
            "description": "New torrents use automatic management (save path from their category).",
            "default": false
          },
          "add_trackers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Trackers added to every new public torrent (never to private ones).",
            "default": []
          },
          "max_ratio": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Share ratio at which the share limit action runs; `null` = no limit.",
            "default": null
          },
          "max_seeding_time": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Seeding time (seconds) at which the share limit action runs; `null` = no limit.",
            "default": null,
            "minimum": 0
          },
          "max_inactive_seeding_time": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Seconds of seeding without traffic after which the share limit action\nruns; `null` = no limit.",
            "default": null,
            "minimum": 0
          },
          "share_limit_action": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ShareLimitAction",
                "description": "What happens when a share limit is reached."
              }
            ],
            "default": "stop"
          },
          "hash_threads": {
            "type": "integer",
            "format": "int32",
            "description": "SHA-1 hashing threads. Applies after a restart.",
            "default": 2,
            "minimum": 0
          },
          "max_open_files": {
            "type": "integer",
            "format": "int32",
            "description": "Content files kept open at once. Applies after a restart.",
            "default": 512,
            "minimum": 0
          },
          "max_checking": {
            "type": "integer",
            "format": "int32",
            "description": "Torrents checked at once. Applies after a restart.",
            "default": 1,
            "minimum": 0
          },
          "piece_extent_affinity": {
            "type": "boolean",
            "description": "Prefer finishing recently started 4 MiB extents. Applies after a restart.",
            "default": true
          },
          "max_concurrent_announces": {
            "type": "integer",
            "format": "int32",
            "description": "Tracker requests in flight at once. Applies after a restart.",
            "default": 32,
            "minimum": 0
          },
          "disk_thread": {
            "type": "boolean",
            "description": "Run disk I/O on its own io_uring thread. Applies after a restart.",
            "default": true
          },
          "zero_copy_send": {
            "type": "boolean",
            "description": "Zero-copy sends for piece payloads. Applies after a restart.",
            "default": false
          },
          "api_session_timeout": {
            "type": "integer",
            "format": "int64",
            "description": "Idle seconds after which an API login session expires.",
            "default": 3600,
            "minimum": 0
          },
          "api_max_auth_failures": {
            "type": "integer",
            "format": "int32",
            "description": "Failed logins from one address before it is banned.",
            "default": 5,
            "minimum": 0
          },
          "api_ban_duration": {
            "type": "integer",
            "format": "int64",
            "description": "Seconds an address stays banned after too many failed logins.",
            "default": 3600,
            "minimum": 0
          },
          "api_bypass_local_auth": {
            "type": "boolean",
            "description": "Loopback clients need no authentication.",
            "default": false
          },
          "api_auth_whitelist": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Address blocks (`10.0.0.0/8`, `fd00::/8`) whose clients need no authentication.",
            "default": []
          },
          "api_allowed_hosts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Host names accepted in the `Host` header (`*` = any, `*.example.com` =\nsubdomains). IP addresses are always accepted.",
            "default": [
              "localhost"
            ]
          },
          "api_csrf_protection": {
            "type": "boolean",
            "description": "Reject state-changing requests from browsers on other origins.",
            "default": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API key from POST /api/v1/auth/api-key."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "urtorrentd_sid",
        "description": "Set by POST /api/v1/auth/login."
      }
    }
  },
  "security": [
    {
      "session": []
    },
    {
      "api_key": []
    }
  ],
...
}